feat: revert token metadata
