+++ packages/token-metadata/lib/index.d.ts
@@ -1,0 +1,3 @@
+export * from './token_metadata';
+export * from './utils';
+//# sourceMappingURL=index.d.ts.map
