--- packages/token-metadata/src/index.ts
@@ -1,2 +1,0 @@
-export * from './utils';
-export * from './token_metadata';
