--- packages/token-metadata/tsconfig.json
@@ -1,10 +1,0 @@
-{
-    "extends": "../../tsconfig",
-    "compilerOptions": {
-        "outDir": "lib",
-        "rootDir": ".",
-        "resolveJsonModule": true,
-        "esModuleInterop": true
-    },
-    "include": ["./src/**/*"],
-}
