--- packages/token-metadata/tslint.json
@@ -1,3 +1,0 @@
-{
-    "extends": ["@0x/tslint-config"]
-}
